+++ src/test/java/blackjack/domain/player/DealerTest.java
@@ -1,7 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.HitFlag;
 import blackjack.domain.card.Deck;
 import blackjack.domain.player.Dealer;
 import org.junit.jupiter.api.DisplayName;
